*Add Screenshot feature, Z button on any GCPad will save a screenshot in apps/wiiflow/ *Removed settings option "Search Mode" (next added button will go here, or will be rearranged) it is no longer needed with the new layout. *Change button layout: -B+UP is now search by previous alphabet letter -B+DOWN is now search by next letter -B+RIGHT is next page -B+LEFT is previous page -B+MINUS is cycle selected partition -B+PLUS is cycle sorting types (NOTE: Using B+UP/DOWN will set sorting to alpha)
*Fix a bug where sorting would crash wiiflow if titles.ini is not present.
*Further cleaning of the ini loading and paths detection.
*Changed some types of some variables (should be more to follow) so they are restricted to acceptable ranges and use less space in memory.

git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@121 43c7226f-5572-e80b-0314-b0280708705f

+++ source/menu/menu_download.cpp
@@ -618,14 +618,14 @@ void CMenu::_textDownload(void)
 	m_btnMgr.setText(m_downloadLblWiiTDB, _t("dl10", L"Please donate\nto WiiTDB.com"));
 }
 
-int CMenu::_versionTxtDownloaderInit(CMenu *m) //Handler to download versions txt file
+s8 CMenu::_versionTxtDownloaderInit(CMenu *m) //Handler to download versions txt file
 {
 	if (!m->m_thrdWorking)
 		return 0;
 	return m->_versionTxtDownloader();
 }
 
-int CMenu::_versionTxtDownloader() // code to download new version txt file
+s8 CMenu::_versionTxtDownloader() // code to download new version txt file
 {
 	block vtxt;
 	char ip[16];
@@ -721,14 +721,14 @@ int CMenu::_versionTxtDownloader() // code to download new version txt file
 	return 0;
 }
 
-int CMenu::_versionDownloaderInit(CMenu *m) //Handler to download new dol
+s8 CMenu::_versionDownloaderInit(CMenu *m) //Handler to download new dol
 {
 	if (!m->m_thrdWorking)
 		return 0;
 	return m->_versionDownloader();
 }
 
-int CMenu::_versionDownloader() // code to download new dol
+s8 CMenu::_versionDownloader() // code to download new dol
 {
 	block newdol;
 	char ip[16];
